Access viaolation exception fix

+++ WinApiNotepadDemo/WinApiWrapper/EnumWindowProcess.cs
@@ -8,6 +8,6 @@
         /// <param name="hWnd">Window handle</param>
         /// <param name="parameter">Caller-defined variable; we use it for a pointer to our list</param>
         /// <returns>True to continue enumerating, false to bail.</returns>
-        public delegate bool EnumWindowsProcess(IntPtr hWnd, int parameter);
+        public delegate bool EnumWindowsProcess(IntPtr hWnd, IntPtr parameter);
     }
 }
